New text on every refresh click
There was a bug that could give the user the same text after clicking refresh. This was because it just gave a random text from the list, including the currently selected one.

I used a filter to remove the currently selected one from a temporary array, and selected the new text from that one.