End game early bug fix
If the user finished the text before the time ran out, the timer would reset to 30 then count down to 29.

I used a set timeout to delay the reseting of the clock to 30, which avoids a state change and stops the clock from going from 30 to 29 after the game.

Although this may not be the most elegant solution, it has solved the problem.